requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.21.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-ibm>=0.0.1
faiss-cpu>=1.7.4
python-dotenv>=0.19.0
ibm-watsonx-ai>=0.1.0
urllib3>=2.0.0
certifi>=2023.7.22
torch>=2.0.0
sentence-transformers>=2.2.2
huggingface-hub>=0.19.4